requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
pillow==10.2.0
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
whitenoise==6.6.0
